refactor: separate useTeleport composable, replace detachable mixin
